Homework 11/pyproject.toml
[tool.poetry]
name = "homework 11"
version = "0.1.0"
description = "Assistant"
authors = ["Garmr <nvova@i.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
marshmallow = "^3.18.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
bcrypt = "^4.0.0"
dotenv = "^0.0.5"
update = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
